pyproject.toml
[project]
name = "ai-interview-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "assemblyai>=0.40.2",
    "bcrypt>=4.3.0",
    "elevenlabs>=1.59.0",
    "fastapi>=0.115.12",
    "google-auth-oauthlib>=1.2.2",
    "google-cloud-speech>=2.32.0",
    "google-cloud-texttospeech>=2.27.0",
    "google-genai>=1.14.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "mem0ai>=0.1.100",
    "numpy>=2.2.6",
    "openai>=1.79.0",
    "passlib>=1.7.4",
    "prisma>=0.15.0",
    "pyaudio>=0.2.14",
    "pydantic-settings>=2.9.1",
    "pydantic[email]>=2.11.4",
    "pyjwt>=2.10.1",
    "pypdf>=5.5.0",
    "python-docx>=1.1.2",
    "python-multipart>=0.0.20",
    "pyttsx3>=2.98",
    "requests>=2.32.3",
    "sounddevice>=0.5.2",
    "soundfile>=0.13.1",
    "uvicorn>=0.34.2",
    "whisper>=1.1.10",
]
